backend-assignment-1/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymongo[srv]>=4.6.1
python-dotenv>=1.0.0
pydantic>=2.5.0
motor>=3.3.2
